Readjusted audio balance for Pizza Time (And More!)
Readjusted the audio balance for It's Pizza Time (event:/music/pizzatime) alongside replacing the Lap 3 theme for that event with "NOTIMETOBOOGIEWOOGIE WIP1"

Also adjusted the rank music states for everyone to be more precise and also to remove the P-Rank music from the rankSP, rankBN and rankPN events, making them use the S-Rank track instead.

I also added in the old Sugary "scrumptious winner" rank tracks as its own event (rankSP Demo) in the chance that timing for the Sugary results screen becomes FMOD related instead of baked in.